+++ dev/src/index.ts
@@ -14,10 +14,16 @@
  * limitations under the License.
  */
 
+<<<<<<< HEAD
 import * as firestore from '@google-cloud/firestore';
 
 import {CallOptions, grpc} from 'google-gax';
 import {Duplex, PassThrough, Transform} from 'stream';
+=======
+import {CallOptions, RetryOptions, Status} from 'google-gax';
+import {Duplex, PassThrough} from 'stream';
+import * as through2 from 'through2';
+>>>>>>> master
 import {URL} from 'url';
 
 import {google} from '../protos/firestore_v1_proto_api';
@@ -53,8 +59,13 @@ import {
   UnaryMethod,
 } from './types';
 import {
+<<<<<<< HEAD
   autoId,
   Deferred,
+=======
+  Deferred,
+  getRetryParams,
+>>>>>>> master
   isPermanentRpcError,
   requestTag,
   wrapError,
@@ -451,19 +462,9 @@ export class Firestore implements firestore.Firestore {
           const grpcModule = this._settings.grpc ?? grpc;
           const sslCreds = grpcModule.credentials.createInsecure();
 
-          // Use user-provided headers, but provide an Authorization header by default
-          // so that connection is recognized as admin in Firestore Emulator. (If for
-          // some reason we're not connecting to the emulator, then this will result in
-          // denials with invalid token, rather than behave like clients not logged in.)
-          const customHeaders = {
-            Authorization: 'Bearer owner',
-            ...this._settings.customHeaders,
-          };
-
           client = new module.exports.v1({
             sslCreds,
             ...this._settings,
-            customHeaders,
           });
         } else {
           client = new module.exports.v1(this._settings);
@@ -1176,6 +1177,20 @@ export class Firestore implements firestore.Firestore {
   async initializeIfNeeded(requestTag: string): Promise<void> {
     this._settingsFrozen = true;
 
+    if (this._settings.ssl === false) {
+      // If SSL is false, we assume that we are talking to the emulator. We
+      // provide an Authorization header by default so that the connection is
+      // recognized as admin in Firestore Emulator. (If for some reason we're
+      // not connecting to the emulator, then this will result in denials with
+      // invalid token, rather than behave like clients not logged in. The user
+      // can then provide their own Authorization header, which will take
+      // precedence).
+      this._settings.customHeaders = {
+        Authorization: 'Bearer owner',
+        ...this._settings.customHeaders,
+      };
+    }
+
     if (this._projectId === undefined) {
       try {
         this._projectId = await this._clientPool.run(requestTag, gapicClient =>
@@ -1203,8 +1218,11 @@ export class Firestore implements firestore.Firestore {
    * Returns GAX call options that set the cloud resource header.
    * @private
    */
-  private createCallOptions(): CallOptions {
-    return {
+  private createCallOptions(
+    methodName: string,
+    retryCodes?: number[]
+  ): CallOptions {
+    const callOptions: CallOptions = {
       otherArgs: {
         headers: {
           [CLOUD_RESOURCE_HEADER]: this.formattedName,
@@ -1212,6 +1230,13 @@ export class Firestore implements firestore.Firestore {
         },
       },
     };
+
+    if (retryCodes) {
+      const retryParams = getRetryParams(methodName);
+      callOptions.retry = new RetryOptions(retryCodes, retryParams);
+    }
+
+    return callOptions;
   }
 
   /**
@@ -1234,7 +1259,7 @@ export class Firestore implements firestore.Firestore {
    * and GAX options.
    * @param requestTag A unique client-assigned identifier for this request.
    * @param func Method returning a Promise than can be retried.
-   * @returns  - A Promise with the function's result if successful within
+   * @returns A Promise with the function's result if successful within
    * `attemptsRemaining`. Otherwise, returns the last rejected Promise.
    */
   private async _retry<T>(
@@ -1262,7 +1287,7 @@ export class Firestore implements firestore.Firestore {
       } catch (err) {
         lastError = err;
 
-        if (isPermanentRpcError(err, methodName, serviceConfig)) {
+        if (isPermanentRpcError(err, methodName)) {
           break;
         }
       }
@@ -1395,14 +1420,17 @@ export class Firestore implements firestore.Firestore {
    * and GAX options.
    * @param request The Protobuf request to send.
    * @param requestTag A unique client-assigned identifier for this request.
+   * @param retryCodes If provided, a custom list of retry codes. If not
+   * provided, retry is based on the behavior as defined in the ServiceConfig.
    * @returns A Promise with the request result.
    */
   request<Req, Resp>(
     methodName: FirestoreUnaryMethod,
     request: Req,
-    requestTag: string
+    requestTag: string,
+    retryCodes?: number[]
   ): Promise<Resp> {
-    const callOptions = this.createCallOptions();
+    const callOptions = this.createCallOptions(methodName, retryCodes);
 
     return this._clientPool.run(requestTag, async gapicClient => {
       try {
@@ -1444,7 +1472,7 @@ export class Firestore implements firestore.Firestore {
     request: {},
     requestTag: string
   ): Promise<Duplex> {
-    const callOptions = this.createCallOptions();
+    const callOptions = this.createCallOptions(methodName);
 
     const bidirectional = methodName === 'listen';